refactor(cmd): repair.go 按子命令拆分 — 1495 行上帝文件 → 7 个按域文件 (#611)
纯移动无逻辑变化：repair.go(203, 共享管线+usage 总览) + repair_duration / repair_merge_status / repair_fragments / repair_delete_by_format / repair_prune(含 reclaim_orphan_merges) / repair_util(共享工具)
